add some generators

+++ src/test/java/personicle/datagen/nosqlcomp/sensoring/Spatial3DPoint.java
@@ -1,11 +1,11 @@
 package personicle.datagen.nosqlcomp.sensoring;
 
 public class Spatial3DPoint {
-    private int x;
-    private int y;
-    private int z;
+    private Double x;
+    private Double y;
+    private Double z;
 
-    public Spatial3DPoint(int x, int y, int z) {
+    public Spatial3DPoint(Double x, Double y, Double z) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -14,27 +14,27 @@ public class Spatial3DPoint {
     public Spatial3DPoint() {
     }
 
-    public int getX() {
+    public Double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(Double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public Double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(Double y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public Double getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(Double z) {
         this.z = z;
     }
 }
